% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LZeroSpikeInference.R
\docType{package}
\name{LZeroSpikeInference}
\alias{LZeroSpikeInference}
\alias{LZeroSpikeInference-package}
\title{LZeroSpikeInference: A package for computing spike train inference with the L0 penalty}
\description{
The LZeroSpikeInference package provides functions for calculting the optimal solution and plotting results. A cross validation #' function is included to aide the user in determining appropriate choices of the tuning parameter and other model parameters.
}
\section{LZeroSpikeInference functions}{

segment
cv
plotSegment
plotCV
}

