% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{optimParams}
\alias{optimParams}
\title{Optimize model parameters based on a fixed tuning parameter (and spike times).}
\usage{
optimParams(params, dat, changePts, penalty, type)
}
\arguments{
\item{params}{model parameters}

\item{dat}{flourscence}

\item{changePts}{locations of changepoints}

\item{penalty}{tuning parameter}

\item{type}{type of model}
}
\description{
Optimize model parameters based on a fixed tuning parameter (and spike times).
}
