% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.cvSpike}
\alias{plot.cvSpike}
\title{Plot mean squared error vs. tuning parameter from the cross-validation output}
\usage{
\method{plot}{cvSpike}(x, ...)
}
\arguments{
\item{x}{output from cross validation procedure}

\item{...}{arguments to be passed to methods}
}
\description{
Plot mean squared error vs. tuning parameter from the cross-validation output
}
