% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.simdata}
\alias{plot.simdata}
\title{Plot simulated data}
\usage{
\method{plot}{simdata}(dat, xlims = NULL)
}
\arguments{
\item{dat}{output data from simulateAR1 or simulateDexp}

\item{xlims}{optional parameter to specify the x-axis limits}
}
\value{
Plot with simulated fluorescence (dark grey circles), calcium concentration (dark green line) and spikes (dark green tick marks on x-axis)
}
\description{
Plot simulated data
}
\examples{

sim <- simulateAR1(n = 500, gam = 0.998, poisMean = 0.009, sd = 0.05, seed = 1)
plot(sim)

sim <- simulateDexp(n = 500, gams = c(0.998, 0.7), poisMean = 0.009, sd = 0.05, seed = 1)
plot(sim)

}
\seealso{
\strong{Estimate spikes:}
\code{\link{estimateSpikes}},
\code{\link{print.estimatedSpikes}},
\code{\link{plot.estimatedSpikes}}.

\strong{Cross validation:}
\code{\link{cv.estimateSpikes}},
\code{\link{print.cvSpike}},
\code{\link{plot.cvSpike}}.

\strong{Simulation:}
\code{\link{simulateAR1}},
\code{\link{simulateDexp}},
\code{\link{plot.simdata}}.
}
