% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotCV}
\alias{plotCV}
\title{Plot mean squared error vs. tuning parameter from the cross-validation output}
\usage{
plotCV(cvOut)
}
\arguments{
\item{cvOut}{output list from cross validation procedure}
}
\description{
Plot mean squared error vs. tuning parameter from the cross-validation output
}
