% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot the solution to an L0 segmentation problem}
\usage{
plotSegmentation(lzeroFit, trueSpikes = NULL, xTimes = NULL,
  spikeTimes = NULL)
}
\arguments{
\item{lzeroFit}{output from running a segmentation}

\item{trueSpikes}{if true spikes are known; this plots them}

\item{xTimes}{times other than standard index}

\item{spikeTimes}{if spike times are known on the same scale as xtimes (in this case no not include trueSpikes)}
}
\description{
Plot the solution to an L0 segmentation problem
}
