% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.cvSpike}
\alias{print.cvSpike}
\title{Print CV results}
\usage{
\method{print}{cvSpike}(x, ...)
}
\arguments{
\item{x}{output from CV}

\item{...}{arguments to be passed to methods}
}
\description{
Print CV results
}
