% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.simdata}
\alias{print.simdata}
\title{Print simulated data}
\usage{
\method{print}{simdata}(x, ...)
}
\arguments{
\item{x}{simulated data}

\item{...}{arguments to be passed to methods}
}
\description{
Print simulated data
}
