% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateAR1}
\alias{simulateAR1}
\title{Simulate fluorescence trace based on simple generative model
y_t = c_t + eps, eps ~ N(0, sd)
c_t = decay * c_{t-1} + s_t
s_t ~ Pois(poisRate)}
\usage{
simulateAR1(n, seed, poisRate, decay, sd)
}
\arguments{
\item{n}{number of timesteps}

\item{seed}{random seed}

\item{poisRate}{Poisson mean spike}

\item{decay}{ar(1) decay rate}

\item{sd}{standard deviation for white noise process}
}
\value{
list of spikes, fluorescence, and calcium concentration
}
\description{
Simulate fluorescence trace based on simple generative model
y_t = c_t + eps, eps ~ N(0, sd)
c_t = decay * c_{t-1} + s_t
s_t ~ Pois(poisRate)
}
\examples{
simulateAR1(n = 100, seed = 1, poisRate = 0.05, decay = 0.998, sd = 0.05)

}
