% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateDexp}
\alias{simulateDexp}
\title{Simulate fluorescence trace based on difference of exponential
calcium dynamics}
\usage{
simulateDexp(n, gams, poisMean, sd, seed)
}
\arguments{
\item{n}{number of timesteps}

\item{gams}{vector (gammaC, gammaD)}

\item{poisMean}{mean for Poisson distributed spikes}

\item{sd}{standard deviation}

\item{seed}{random seed}
}
\value{
list of spikes, fluorescence, and calcium concentration
}
\description{
Simulate fluorescence trace based on difference of exponential
calcium dynamics
}
\details{
Simulate fluorescence trace based on difference of exponential
calcium dynamics

y_t =  r_t + eps, eps ~ N(0, sd)

c_t = gammaC * c_{t-1} + s_t

d_t = gammaD * d_{t-1} + s_t

r_t = (c_t - d_t)

s_t ~ Pois(poisMean)
}
\examples{
sim <- simulateDexp(n = 500, gams = c(0.998, 0.7), poisMean = 0.009, sd = 0.05, seed = 1)
plot(sim)
}
\seealso{
\strong{Estimate spikes:}
\code{\link{estimateSpikes}},
\code{\link{print.estimatedSpikes}},
\code{\link{plot.estimatedSpikes}}.

\strong{Cross validation:}
\code{\link{cv.estimateSpikes}},
\code{\link{print.cvSpike}},
\code{\link{plot.cvSpike}}.

\strong{Simulation:}
\code{\link{simulateAR1}},
\code{\link{simulateDexp}},
\code{\link{plot.simdata}}.
}
