% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.cvSpike}
\alias{plot.cvSpike}
\title{Plot mean squared error vs. tuning parameter from the cross-validation output}
\usage{
\method{plot}{cvSpike}(cvOut)
}
\arguments{
\item{cvOut}{output from cross validation procedure}
}
\description{
Plot mean squared error vs. tuning parameter from the cross-validation output
}
