% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.estimatedSpikes}
\alias{plot.estimatedSpikes}
\title{Plot the solution to an L0 segmentation problem}
\usage{
\method{plot}{estimatedSpikes}(x, xlims = NULL, ...)
}
\arguments{
\item{x}{output from running estimatedSpikes}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{arguments to be passed to methods}
}
\description{
Plot the solution to an L0 segmentation problem
}
\seealso{
\strong{Estimate spikes:}
\code{\link{estimateSpikes}},
\code{\link{print.estimatedSpikes}},
\code{\link{plot.estimatedSpikes}}.

\strong{Cross validation:}
\code{\link{cv.estimateSpikes}},
\code{\link{print.cvSpike}},
\code{\link{plot.cvSpike}}.

\strong{Simulation:}
\code{\link{simulateAR1}},
\code{\link{plot.simdata}}.
}
