% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot the solution to an L0 segmentation problem}
\usage{
plotSegmentation(lzeroFit, trueSpikes = NULL, xTimes = NULL,
  spikeTimes = NULL)
}
\arguments{
\item{lzeroFit}{output from running a segmentation}

\item{trueSpikes}{if true spikes are known and provided this adds tick marks at specified index locations}

\item{xTimes}{timeseries corresponding to the time that each index of the fluorescence data correspond to. Include to plot the x-axis on this scale}

\item{spikeTimes}{times, on the same scale as xTimes, that spikes occur in the true (ground truth) data}
}
\description{
Plot the solution to an L0 segmentation problem
}
