% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.estimatedSpikes}
\alias{print.estimatedSpikes}
\title{Print estimated spikes}
\usage{
\method{print}{estimatedSpikes}(lzeroFit)
}
\arguments{
\item{lzeroFit}{estimated spikes}
}
\description{
Print estimated spikes
}
