% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segment}
\alias{segment}
\title{Segments based on type of model and penalty}
\usage{
segment(dat, params, penalty, type = c("ar1", "intercept", "dexp"),
  calcFittedValues = TRUE)
}
\arguments{
\item{dat}{noisy flourscence data}

\item{params}{model parameters. For the AR(1) and AR(1) intercept models this is the scalar decay parameter. The difference of exponentials model this is a dataframe, for example \code{params <- data.frame(gammaC = 0.98, gammaD = 0.818)}}

\item{penalty}{tuning parameter}

\item{type}{type of model, must be one of AR(1) 'ar1', AR(1) with intercept 'intercept', or difference of exponentials 'dexp'}

\item{calcFittedValues}{parameter to determine whether to calculate fitted values. FALSE for example in cross validation}
}
\value{
Returns a list with elements:

\code{changePts} the set of changepoints

\code{spikes} the set of spikes

\code{fittedValues} estimated calcium concentration

\code{dat} the data that was used to fit the model
}
\description{
Segments based on type of model and penalty
}
\examples{

sampleData <- simulateAR1(n = 500, seed = 1, poisRate = 0.01, decay = 0.998, sd = 0.15)

# AR(1) model

fit <- segment(sampleData$fl, params = 0.998, penalty = 1, type = "ar1")
plotSegmentation(fit, trueSpikes = sampleData$spikes)

# AR(1) + intercept model

fit <- segment(sampleData$fl, params = 0.998, penalty = 1, type = "intercept")
plotSegmentation(fit, trueSpikes = sampleData$spikes)

# Difference of exponentials model

params <- data.frame(gammaC = 0.998, gammaD = 0.82)
fit <- segment(sampleData$fl, params, penalty = 1, type = "dexp")
plotSegmentation(fit, trueSpikes = sampleData$spikes)

}
