% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateAR1}
\alias{simulateAR1}
\title{Simulate flourscence trace based on generative model in paper}
\usage{
simulateAR1(n, seed, poisRate, decay, sd)
}
\arguments{
\item{n}{number of timesteps}

\item{seed}{random seed}

\item{poisRate}{Poisson mean spike}

\item{decay}{ar(1) decay rate}

\item{sd}{standard deviation for white noise process}
}
\value{
list of spikes, flourscence, and concentration
}
\description{
Simulate flourscence trace based on generative model in paper
}
\examples{
simulateAR1(n = 100, seed = 1, poisRate = 0.05, decay = 0.998, sd = 0.05)

}
